Rewrite to ES6 and refactor event handling module
